+++ server/internal/orm/model/role.gen.go
@@ -12,13 +12,13 @@ const TableNameRole = "role"
 
 // Role 用户角色
 type Role struct {
-	ID          int32  `gorm:"column:id;primaryKey;autoIncrement:true" json:"id"`
-	AppID       int32  `gorm:"column:app_id;not null;comment:应用ID" json:"app_id"`                          // 应用ID
+	ID          int64  `gorm:"column:id;primaryKey;autoIncrement:true" json:"id"`
+	AppID       int64  `gorm:"column:app_id;not null;comment:应用ID" json:"app_id"`                          // 应用ID
 	Name        string `gorm:"column:name;not null;comment:角色名称" json:"name"`                              // 角色名称
 	Rules       string `gorm:"column:rules;not null;comment:权限ID ,分割a" json:"rules"`                       // 权限ID ,分割a
 	RulesCheckd string `gorm:"column:rules_checkd;not null;comment:权限树选中的字节点ID" json:"rules_checkd"`       // 权限树选中的字节点ID
-	CreateTime  int32  `gorm:"column:create_time;not null;autoCreateTime;comment:创建时间" json:"create_time"` // 创建时间
-	UpdateTime  int32  `gorm:"column:update_time;not null;autoUpdateTime;comment:更新时间" json:"update_time"` // 更新时间
+	CreateTime  int64  `gorm:"column:create_time;not null;autoCreateTime;comment:创建时间" json:"create_time"` // 创建时间
+	UpdateTime  int64  `gorm:"column:update_time;not null;autoUpdateTime;comment:更新时间" json:"update_time"` // 更新时间
 	Remark      string `gorm:"column:remark;not null;comment:简介" json:"remark"`                            // 简介
 	Status      int32  `gorm:"column:status;not null;default:1;comment:状态" json:"status"`                  // 状态
 	Sort        int32  `gorm:"column:sort;not null;comment:排序" json:"sort"`                                // 排序
